refactoring the eval conditions
