+++ scala-backend/app/saladbar/src/main/scala/Bop.scala
@@ -8,6 +8,22 @@ package saladbar
   */
 sealed trait Bop {
     override def toString: String
+    def checkBop1[A](v1: Value)(sc: () => A)(fc: () => A): A = this match {
+        case And | Or =>  v1 match {
+            case B(_) => sc()
+            case _ => fc()
+        }
+        case Eq | Neq | Eqq | Neqq => sc()
+        case Gt | Geq | Lt | Leq | Plus => v1 match {
+            case N(_) | S(_) => sc()
+            case _ => fc()
+        }
+        case Minus | Times | Div => v1 match {
+            case N(_) => sc()
+            case _ => fc()
+        }
+        case _ => ???
+    }
 }
 
 
@@ -28,6 +44,11 @@ case object Plus extends Bop {
 
 case object Times extends Bop{
     override def toString: String = "*"
+}
+
+
+case object Div extends Bop{
+    override def toString: String = "/"
 }
