+++ scala-backend/app/saladbar/src/main/scala/DynamicScopingError.scala
@@ -1,0 +1,18 @@
+package saladbar
+
+
+/**
+  * DynamicScopingError
+  * 
+  * Used in DynamicScope
+  *
+  * @param msg
+  */
+case class DynamicScopeError(msg: String) extends Exception{
+
+
+  override def toString: String = s"DYNAMIC SCOPING ERROR: $msg"
+
+
+}
+
