+++ scala-backend/app/saladbar/src/main/scala/LazyEagerCondition.scala
@@ -1,0 +1,25 @@
+package saladbar
+
+
+/**
+  * LazyEagerCondition
+  */
+abstract class LazyEagerCondition {
+
+
+    /**
+      * check
+      * 
+      * confirm if $e1 is ready for substitution subject to the
+      * scoping condition. if so, execute $sc, else execute $fc
+      *
+      * @param e1
+      * @param sc
+      * @param fc
+      * @return
+      */
+    def check[A](e1: Expr)(sc: () => A)(fc: () => A): A
+
+
+}
+
