+++ scala-backend/app/saladbar/src/main/scala/LexicalScopeError.scala
@@ -1,0 +1,13 @@
+package saladbar
+
+
+/**
+  * LexicalScopingError
+  * 
+  * Used in LexicalScope
+  *
+  * @param msg
+  */
+case class LexicalScopeError(msg: String) extends Exception{
+  override def toString: String = s"LEXICAL SCOPING ERROR: $msg"
+}
