+++ scala-backend/app/saladbar/src/main/scala/NoConversionsError.scala
@@ -1,0 +1,18 @@
+package saladbar
+
+
+/**
+  * NoConversionsError
+  * 
+  * Used in NoConversions
+  *
+  * @param msg
+  */
+case class NoConversionsError(msg: String) extends Exception {
+
+
+  override def toString: String = s"NO CONVERSIONS: $msg"
+
+
+}
+
