+++ scala-backend/app/saladbar/src/main/scala/ScopingCondition.scala
@@ -1,0 +1,13 @@
+package saladbar
+
+
+/**
+  * ScopingCondition
+  */
+abstract class ScopingCondition {
+  def substFunctions[A](evalConditions: EvalConditions, 
+      e: Expr,
+      x: String,
+      esub: Expr)(sc: Expr => A):A 
+}
+
